refactor both generators to follow similar code and write methods
