+++ similar_listings/package.json
@@ -14,10 +14,8 @@
     "react-dev": "webpack -d --watch",
     "react-prod": "webpack -d",
     "test": "jest",
-    "mongo-seed": "mongoimport --database listing --collection listings --file data/data.json",
-    "data": "node data/largeSeed.js",
-    "rm-data": "rm data/data.json",
-    "start-db": "mongod"
+    "data-csv": "node data/CSVgen.js",
+    "data-json": "node data/JSONgen.js",
   },
   "dependencies": {
     "axios": "^0.18.0",
